Fixed install process and removed templates folder support

+++ install.php
@@ -27,7 +27,7 @@ file_exists($dumboSystemPath) || mkdir($dumboSystemPath, 0777, TRUE);
 
 $d = dir($path);
 while (false !== ($entry = $d->read())) {
-   if($entry != '.' && $entry != '..' && $entry != 'install.php' && $entry != 'src' && $entry != 'lib' && $entry != '.git' && $entry != '.gitignore'){
+   if($entry != '.' && $entry != '..' && $entry != 'install.php' && !is_dir($entry) && $entry != 'src' && $entry != 'lib' && $entry != '.git' && $entry != '.gitignore'){
    		echo 'copying '.$path.'/'.$entry.' to '.$dumboSystemPath.'/'.$entry.PHP_EOL;
    		file_exists($dumboSystemPath.'/'.$entry) && unlink($dumboSystemPath.'/'.$entry);
    		copy($path.'/'.$entry, $dumboSystemPath.'/'.$entry) or die('Could not copy file.');
